Updated readme and added a story for @jmrenner
So btw git add patch mode aka git add -p allows you to go through
diff by diff or in git terms glob by glob and deside what you want
to do with it! So great got not commit logging and debugging code
but not lossing it as well. Along with 1000000 other uses. Check it
out.